requirements-clean.txt
anyio==4.13.0
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
base58==2.1.1
bcrypt==3.2.2
beautifulsoup4==4.14.3
blinker==1.7.0
boto3==1.34.46
botocore==1.34.46
certifi==2026.4.22
cffi==2.0.0
chardet==5.2.0
click==8.1.6
cloud-init==25.3
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
construct==2.10.70
construct-typing==0.7.0
cryptography==41.0.7
curl_cffi==0.15.0
dbus-python==1.3.2
distro==1.9.0
distro-info==1.7+build1
fail2ban==1.0.2
frozendict==2.4.7
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Jinja2==3.1.2
jmespath==1.0.1
jsonalias==0.1.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
krakenex==2.2.2
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mnemonic==0.21
multitasking==0.0.13
netifaces==0.11.0
numpy==2.4.3
oandapyV20==0.7.2
oauthlib==3.2.2
packaging==24.0
pandas==3.0.1
peewee==4.0.4
pexpect==4.9.0
platformdirs==4.9.6
protobuf==7.34.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyasyncore==1.0.2
pycparser==3.0
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
pyinotify==0.9.6
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5.2
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.2.2
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
s3transfer==0.10.1
schedule==1.2.1
service-identity==24.1.0
setuptools==68.1.2
six==1.16.0
solana==0.36.11
solders==0.27.1
sos==4.10.2
soupsieve==2.8.3
ssh-import-id==5.11
systemd-python==235
ta==0.11.0
Twisted==24.3.0
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
wadllib==1.3.6
websockets==15.0.1
wheel==0.42.0
xkit==0.0.0
yfinance==1.3.0
zope.interface==6.1
